Move projectiles into their own directory.
