+++ src/main/java/twilightforest/entity/EntityTFSlimeProjectile.java
@@ -13,6 +13,7 @@ import net.minecraft.util.math.RayTraceResult;
 import net.minecraft.world.World;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
+import twilightforest.entity.projectile.EntityTFThrowable;
 
 public class EntityTFSlimeProjectile extends EntityTFThrowable implements IRendersAsItem {
