+++ src/main/java/twilightforest/entity/projectile/EntityTFTwilightWandBolt.java
@@ -1,4 +1,4 @@
-package twilightforest.entity;
+package twilightforest.entity.projectile;
 
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.IRendersAsItem;
