+++ src/main/java/twilightforest/entity/projectile/ITFProjectile.java
@@ -1,4 +1,4 @@
-package twilightforest.entity;
+package twilightforest.entity.projectile;
 
 import net.minecraft.entity.IProjectile;
